Don't expose server to the network
Previously, Mockito bound its server to 0.0.0.0, which exposed the
server to every host on the network. This _probably_ isn't the correct
behavior. It also generated several annoying firewall popups per test
run on my machine (using the default macOS firewall settings, AFAIK).

This commit changes the server to run on 127.0.0.1, so it's not
accessible to other machines.