Preserve response header order
I came across some intermittent test failures when working on some code
that serializes HTTP responses. This seemed like it should be
deterministic, so this was very surprising.

It turned out that the problem came from Mockito using a `HashMap` to
store response headers. Switching this to an array, and explcitly
specifying a static `Date` header, should be enough to make Mockito
responses entirely deterministic, which is especially desirable in a
library focused on testing.