Remove unused PartialEq derive
The PartialEq trait was derived for Mock. This implied the need for
Response to implement PartialEq, and, in turn, for Body to implment it
aswell. Clippy complains about the comparison between closures in the
arguments of the Body::Fn enum, that are non-comparable.
The solution is to remove te unneeded PartialEq trait derivation from
Mock and from Response, therefore lifting the bound to implement it for
Body.