Change implementation of PartialEq for Response
The Body enum is a clever way to implement the creation of the response
body either as a static binary (the Bytes variant), or as the result of
a closure (the Fn variant). Unfortunately, two closures, even if
identical, bear different types. They are trait objects and are
represented as a couple of pointers, (data, vtable). The first holds a
reference to the data, the second to a virtual table (created at
runtime) that references the function to be called on that data.
In this case, it should be safe to compare only the (reference to the)
data, and not the function.