Handle httpparse error
When httpparse fails to parse a request, it will return an error. For
instance when a header is invalid, httparse will return error
`Error::HeaderName`.
The current logic was to mark the error and re-enter the loop.
When this happens, the buffer may be empty and the test will block
indefinitely.

This change marks the request as `parsed` when httparse returns an error
parsing the content so this can be returned back to the client.